+++ great_international/serializers.py
@@ -8,6 +8,7 @@ from core.serializers import (
     FormPageSerializerMetaclass,
     ParentPageSerializerHelper
     )
+from invest.models import HighPotentialOpportunityDetailPage
 
 from .models import (
     InternationalArticlePage,
@@ -18,7 +19,13 @@ from .models import (
     InternationalSectorPage,
     InternationalEUExitFormPage,
     CapitalInvestOpportunityPage,
-    CapitalInvestRegionalSectorPage)
+    CapitalInvestRegionalSectorPage,
+    InvestHighPotentialOpportunityFormPage,
+    InvestHighPotentialOpportunityDetailPage)
+
+
+ONE_TO_SIX_WORDS = ['one', 'two', 'three', 'four', 'five', 'six']
+ONE_TO_SEVEN_WORDS = ONE_TO_SIX_WORDS + ['seven']
 
 
 class SectionThreeSubsectionProxyDataWrapper:
@@ -1090,3 +1097,303 @@ class CapitalInvestOpportunityPageSerializer(PageWithRelatedPagesSerializer):
 
 class MinimalPageSerializer(BasePageSerializer):
     hero_title = serializers.CharField(max_length=255)
+
+
+# Invest seralizers
+
+class SubsectionProxyDataWrapper:
+
+    def __init__(self, instance, suffix):
+        self.suffix = suffix
+        self.instance = instance
+
+    @property
+    def title(self):
+        return getattr(self.instance, f'subsection_title_{self.suffix}')
+
+    @property
+    def content(self):
+        return getattr(self.instance, f'subsection_content_{self.suffix}')
+
+    @property
+    def map(self):
+        return getattr(self.instance, f'subsection_map_{self.suffix}', None)
+
+
+class HowWeHelpProxyDataWrapper:
+
+    def __init__(self, instance, suffix):
+        self.suffix = suffix
+        self.instance = instance
+
+    @property
+    def text(self):
+        return getattr(self.instance, f'how_we_help_text_{self.suffix}')
+
+    @property
+    def icon(self):
+        return getattr(
+            self.instance,
+            f'how_we_help_icon_{self.suffix}',
+            None
+        )
+
+
+class SubsectionSerializer(serializers.Serializer):
+    title = serializers.CharField(max_length=255)
+    content = core_fields.MarkdownToHTMLField()
+    map = wagtail_fields.ImageRenditionField(
+        'original',
+        allow_null=True
+    )
+
+
+class HowWeHelpSerializer(serializers.Serializer):
+    text = serializers.CharField(max_length=255)
+    icon = wagtail_fields.ImageRenditionField(
+        'original',
+        allow_null=True
+    )
+
+
+class InvestInternationalHomePageSerializer(BasePageSerializer):
+    breadcrumbs_label = serializers.CharField(max_length=50)
+    heading = serializers.CharField(max_length=255)
+    sub_heading = serializers.CharField(max_length=255)
+    hero_call_to_action_text = serializers.CharField(max_length=255)
+    hero_call_to_action_url = serializers.CharField(max_length=255)
+    hero_image = wagtail_fields.ImageRenditionField('original')
+    benefits_section_title = serializers.CharField(max_length=255)
+    benefits_section_intro = serializers.CharField(max_length=255)
+    benefits_section_content = core_fields.MarkdownToHTMLField()
+    benefits_section_img = wagtail_fields.ImageRenditionField('fill-640x360')
+    capital_invest_section_title = serializers.CharField(max_length=255)
+    capital_invest_section_content = core_fields.MarkdownToHTMLField()
+    capital_invest_section_image = wagtail_fields.ImageRenditionField(
+        'fill-640x360')
+    eu_exit_section_title = serializers.CharField(max_length=255)
+    eu_exit_section_content = core_fields.MarkdownToHTMLField()
+    eu_exit_section_call_to_action_text = serializers.CharField(max_length=255)
+    eu_exit_section_call_to_action_url = serializers.CharField(max_length=255)
+    eu_exit_section_img = wagtail_fields.ImageRenditionField('original')
+    subsections = serializers.SerializerMethodField()
+    sector_title = serializers.CharField(max_length=255)
+    sector_intro = serializers.CharField(max_length=255)
+    hpo_title = serializers.CharField(max_length=255)
+    hpo_intro = serializers.CharField(max_length=255)
+    sector_button_text = serializers.CharField(max_length=255)
+    sector_button_url = serializers.CharField(max_length=255)
+    setup_guide_title = serializers.CharField(max_length=255)
+    setup_guide_content = core_fields.MarkdownToHTMLField()
+    setup_guide_img = wagtail_fields.ImageRenditionField('fill-640x360')
+    setup_guide_call_to_action_url = serializers.CharField(max_length=255)
+    setup_guide_lead_in = serializers.CharField(
+        max_length=255,
+        allow_null=True
+    )
+    isd_section_image = wagtail_fields.ImageRenditionField('fill-640x360')
+    isd_section_title = serializers.CharField(max_length=255)
+    isd_section_text = core_fields.MarkdownToHTMLField(max_length=255)
+    how_we_help_title = serializers.CharField(max_length=255)
+    how_we_help_lead_in = serializers.CharField(max_length=255)
+    how_we_help = serializers.SerializerMethodField()
+    contact_section_title = serializers.CharField(max_length=255)
+    contact_section_content = serializers.CharField(max_length=255)
+    contact_section_call_to_action_text = serializers.CharField(max_length=255)
+    contact_section_call_to_action_url = serializers.CharField(max_length=255)
+    sectors = serializers.SerializerMethodField()
+    high_potential_opportunities = serializers.SerializerMethodField()
+    guides = serializers.SerializerMethodField()
+
+    def get_how_we_help(self, instance):
+        data = [
+            HowWeHelpProxyDataWrapper(instance=instance, suffix=num)
+            for num in ONE_TO_SIX_WORDS
+        ]
+        serializer = HowWeHelpSerializer(data, many=True)
+        return serializer.data
+
+    def get_subsections(self, instance):
+        data = [
+            SubsectionProxyDataWrapper(instance=instance, suffix=num)
+            for num in ONE_TO_SEVEN_WORDS
+        ]
+        serializer = SubsectionSerializer(data, many=True)
+        return serializer.data
+
+    def get_sectors(self, instance):
+        queryset = InternationalSectorPage.objects.all().filter(
+            featured=True
+        ).live().order_by('heading')
+        serializer = InternationalSectorPageSerializer(
+            queryset,
+            many=True,
+            allow_null=True,
+            context=self.context
+        )
+        return serializer.data
+
+    def get_high_potential_opportunities(self, instance):
+        from .models import InvestHighPotentialOpportunityDetailPage
+        queryset = InvestHighPotentialOpportunityDetailPage.objects.all(
+            ).filter(
+            featured=True
+        ).live().order_by('heading')
+        serializer = InvestHighPotentialOpportunityDetailPageSerializer(
+            queryset,
+            many=True,
+            allow_null=True,
+            context=self.context
+        )
+        return serializer.data
+
+
+class InvestHighPotentialOpportunityDetailPageBaseSerializer(
+    BasePageSerializer
+):
+    breadcrumbs_label = serializers.CharField(max_length=50)
+    heading = serializers.CharField(max_length=255)
+    hero_image = wagtail_fields.ImageRenditionField('original')
+    description = serializers.CharField(max_length=255)
+    featured = serializers.BooleanField()
+    contact_proposition = core_fields.MarkdownToHTMLField()
+    contact_button = serializers.CharField(max_length=500)
+    proposition_one = core_fields.MarkdownToHTMLField()
+    proposition_one_image = wagtail_fields.ImageRenditionField('original')
+    proposition_one_video = core_fields.VideoField()
+    opportunity_list_title = serializers.CharField()
+    opportunity_list_item_one = core_fields.MarkdownToHTMLField()
+    opportunity_list_item_two = core_fields.MarkdownToHTMLField()
+    opportunity_list_item_three = core_fields.MarkdownToHTMLField()
+    opportunity_list_image = wagtail_fields.ImageRenditionField('original')
+    proposition_two = core_fields.MarkdownToHTMLField()
+    proposition_two_list_item_one = core_fields.MarkdownToHTMLField()
+    proposition_two_list_item_two = core_fields.MarkdownToHTMLField()
+    proposition_two_list_item_three = core_fields.MarkdownToHTMLField()
+    proposition_two_image = wagtail_fields.ImageRenditionField('original')
+    proposition_two_video = core_fields.VideoField()
+    competitive_advantages_title = serializers.CharField()
+    competitive_advantages_list_item_one = core_fields.MarkdownToHTMLField()
+    competitive_advantages_list_item_one_icon = \
+        wagtail_fields.ImageRenditionField('original')
+    competitive_advantages_list_item_two = core_fields.MarkdownToHTMLField()
+    competitive_advantages_list_item_two_icon = \
+        wagtail_fields.ImageRenditionField('original')
+    competitive_advantages_list_item_three = core_fields.MarkdownToHTMLField()
+    competitive_advantages_list_item_three_icon = \
+        wagtail_fields.ImageRenditionField('original')
+    testimonial = core_fields.MarkdownToHTMLField()
+    testimonial_background = wagtail_fields.ImageRenditionField('original')
+    companies_list_text = core_fields.MarkdownToHTMLField()
+    companies_list_item_image_one = wagtail_fields.ImageRenditionField(
+        'original'
+    )
+    companies_list_item_image_two = wagtail_fields.ImageRenditionField(
+        'original'
+    )
+    companies_list_item_image_three = wagtail_fields.ImageRenditionField(
+        'original'
+    )
+    companies_list_item_image_four = wagtail_fields.ImageRenditionField(
+        'original'
+    )
+    companies_list_item_image_five = wagtail_fields.ImageRenditionField(
+        'original'
+    )
+    companies_list_item_image_six = wagtail_fields.ImageRenditionField(
+        'original'
+    )
+    companies_list_item_image_seven = wagtail_fields.ImageRenditionField(
+        'original'
+    )
+    companies_list_item_image_eight = wagtail_fields.ImageRenditionField(
+        'original'
+    )
+    case_study_list_title = serializers.CharField()
+    case_study_one_text = core_fields.MarkdownToHTMLField()
+    case_study_one_image = wagtail_fields.ImageRenditionField('original')
+    case_study_two_text = core_fields.MarkdownToHTMLField()
+    case_study_two_image = wagtail_fields.ImageRenditionField('original')
+    case_study_three_text = core_fields.MarkdownToHTMLField()
+    case_study_three_image = wagtail_fields.ImageRenditionField('original')
+    case_study_four_text = core_fields.MarkdownToHTMLField()
+    case_study_four_image = wagtail_fields.ImageRenditionField('original')
+    other_opportunities_title = serializers.CharField()
+    pdf_document = core_fields.DocumentURLField()
+    summary_image = wagtail_fields.ImageRenditionField('original')
+
+
+class InvestHighPotentialOpportunityDetailPageSerializer(
+    InvestHighPotentialOpportunityDetailPageBaseSerializer
+):
+    other_opportunities = serializers.SerializerMethodField()
+
+    def get_other_opportunities(self, instance):
+        queryset = (
+            InvestHighPotentialOpportunityDetailPage.objects.all()
+            .live()
+            .order_by('heading')
+            .exclude(slug=instance.slug)
+        )
+        serializer = InvestHighPotentialOpportunityDetailPageBaseSerializer(
+            queryset,
+            many=True,
+            allow_null=True,
+            context=self.context
+        )
+        return serializer.data
+
+
+class InvestHighPotentialOpportunityFormPageSerializer(BasePageSerializer,
+    metaclass=FormPageSerializerMetaclass):
+    class Meta:
+        model_class = InvestHighPotentialOpportunityFormPage
+
+    heading = serializers.CharField(max_length=255)
+    sub_heading = serializers.CharField(max_length=255)
+    breadcrumbs_label = serializers.CharField(max_length=50)
+    opportunity_list = serializers.SerializerMethodField()
+
+    def get_opportunity_list(self, instance):
+        queryset = (
+            HighPotentialOpportunityDetailPage.objects.all()
+            .live()
+            .order_by('heading')
+            .exclude(slug=instance.slug)
+        )
+        serializer = InvestHighPotentialOpportunityDetailPageSerializer(
+            queryset,
+            many=True,
+            allow_null=True,
+            context=self.context
+        )
+        return serializer.data
+
+
+class InvestHighPotentialOpportunityFormSuccessPageSerializer(
+    BasePageSerializer
+):
+    breadcrumbs_label = serializers.CharField(max_length=50)
+    heading = serializers.CharField()
+    sub_heading = serializers.CharField()
+    next_steps_title = serializers.CharField()
+    next_steps_body = serializers.CharField()
+    documents_title = serializers.CharField()
+    documents_body = serializers.CharField()
+
+    opportunity_list = serializers.SerializerMethodField()
+
+    def get_opportunity_list(self, instance):
+        queryset = (
+            InvestHighPotentialOpportunityDetailPage.objects.all()
+            .live()
+            .order_by('heading')
+            .exclude(slug=instance.slug)
+        )
+        serializer = InvestHighPotentialOpportunityDetailPageSerializer(
+            queryset,
+            many=True,
+            allow_null=True,
+            context=self.context
+        )
+        return serializer.data